// src/Components/lib/types.ts
export interface ContentType {
    id: string
    title: string
    description: string
    thumbnail: string
    coverImage: string
    videoUrl: string
    releaseYear: number
    duration: string
    rating: string
    genres: string[]
    cast: string[]
    matchPercentage: number
    isTrending: boolean
    isNewRelease: boolean
  }
  
  